Added Gips, Player death, minor fixes
Gips are now added in the game, though without any AI or animations.
When a player runs out of health, the GameManager now resets the scene.
While not currently useful, various functionality of the player turns
off while the player is dead - this will be useful once the player does
not instantly respawn. An unnecessary variable was removed from
SwordSlash.